requirements.txt
begins
pillow